remove Function Node
