+++ fmriprep/workflows/anatomical.py
@@ -26,6 +26,7 @@ from niworkflows.interfaces.segmentation import FASTRPT, ReconAllRPT
 
 from ..interfaces import DerivativesDataSink, StructuralReference, MakeMidthickness
 from ..interfaces.bids import BIDSInfo
+from ..interfaces.surf import NormalizeSurf
 from ..interfaces.images import ConformSeries
 from ..interfaces.reports import AnatomicalSummary
 from ..utils.misc import fix_multi_T1w_source_name, add_suffix
@@ -82,7 +83,7 @@ def init_anat_preproc_wf(skull_strip_ants, output_spaces, template, debug, frees
     # 4. Segmentation
     t1_seg = pe.Node(FASTRPT(generate_report=True, segments=True,
                              no_bias=True, probability_maps=True),
-                     name='t1_seg')
+                     name='t1_seg', estimated_memory_gb=3)
 
     # 5. Spatial normalization (T1w to MNI registration)
     t1_2_mni = pe.Node(
@@ -92,11 +93,9 @@ def init_anat_preproc_wf(skull_strip_ants, output_spaces, template, debug, frees
             num_threads=omp_nthreads,
             flavor='testing' if debug else 'precise',
         ),
-        name='t1_2_mni'
+        name='t1_2_mni',
+        n_procs=omp_nthreads
     )
-    # should not be necessary but does not hurt - make sure the multiproc
-    # scheduler knows the resource limits
-    t1_2_mni.interface.num_threads = omp_nthreads
 
     # Resample the brain mask and the tissue probability maps into mni space
     mni_mask = pe.Node(
@@ -131,7 +130,8 @@ def init_anat_preproc_wf(skull_strip_ants, output_spaces, template, debug, frees
         (t1_seg, outputnode, [('tissue_class_map', 't1_seg'),
                               ('probability_maps', 't1_tpms')]),
         (inputnode, summary, [('t1w', 't1w')]),
-        ])
+    ])
+
     if 'template' in output_spaces:
         template_str = nid.TEMPLATE_MAP[template]
         ref_img = op.join(nid.get_dataset(template_str), '1mm_T1.nii.gz')
@@ -179,7 +179,7 @@ def init_anat_preproc_wf(skull_strip_ants, output_spaces, template, debug, frees
                 ('outputnode.subject_id', 'subject_id'),
                 ('outputnode.fs_2_t1_transform', 'fs_2_t1_transform'),
                 ('outputnode.surfaces', 'surfaces')]),
-            ])
+        ])
 
     anat_reports_wf = init_anat_reports_wf(
         reportlets_dir=reportlets_dir, skull_strip_ants=skull_strip_ants,
@@ -190,7 +190,7 @@ def init_anat_preproc_wf(skull_strip_ants, output_spaces, template, debug, frees
         (summary, anat_reports_wf, [('out_report', 'inputnode.summary_report')]),
         (t1_conform, anat_reports_wf, [('out_report', 'inputnode.t1_conform_report')]),
         (t1_seg, anat_reports_wf, [('out_report', 'inputnode.t1_seg_report')]),
-        ])
+    ])
 
     if skull_strip_ants:
         workflow.connect([
@@ -226,8 +226,8 @@ def init_anat_preproc_wf(skull_strip_ants, output_spaces, template, debug, frees
             ('mni_seg', 'inputnode.mni_seg'),
             ('mni_tpms', 'inputnode.mni_tpms'),
             ('surfaces', 'inputnode.surfaces'),
-            ]),
-        ])
+        ]),
+    ])
 
     return workflow
 
@@ -246,11 +246,7 @@ def init_skullstrip_ants_wf(debug, omp_nthreads, name='skullstrip_ants_wf'):
         dimension=3, use_floatingpoint_precision=1,
         debug=debug, generate_report=True,
         num_threads=omp_nthreads, keep_temporary_files=1),
-        name='t1_skull_strip')
-
-    # should not be necesssary byt does not hurt - make sure the multiproc
-    # scheduler knows the resource limits
-    t1_skull_strip.interface.num_threads = omp_nthreads
+        name='t1_skull_strip', n_procs=omp_nthreads)
 
     t1_skull_strip.inputs.brain_template = op.join(
         get_ants_oasis_template_ras(),
@@ -321,9 +317,9 @@ def init_surface_recon_wf(omp_nthreads, hires, name='surface_recon_wf'):
             directive='autorecon1',
             flags='-noskullstrip',
             openmp=omp_nthreads),
-        name='autorecon1')
+        name='autorecon1',
+        n_procs=omp_nthreads)
     autorecon1.interface._can_resume = False
-    autorecon1.interface.num_threads = omp_nthreads
 
     def inject_skullstripped(subjects_dir, subject_id, skullstripped):
         import os
@@ -392,7 +388,7 @@ def init_surface_recon_wf(omp_nthreads, hires, name='surface_recon_wf'):
                                            ('outputnode.out_report', 'out_report')]),
         (gifti_surface_wf, outputnode, [('outputnode.surfaces', 'surfaces')]),
         (fs_transform, outputnode, [('fsl_file', 'fs_2_t1_transform')]),
-        ])
+    ])
 
     return workflow
 
@@ -458,12 +454,17 @@ def init_autorecon_resume_wf(omp_nthreads, name='autorecon_resume_wf'):
         (autorecon3, recon_report, [(('subjects_dir', _dedup), 'subjects_dir'),
                                     (('subject_id', _dedup), 'subject_id')]),
         (recon_report, outputnode, [('out_report', 'out_report')]),
-        ])
+    ])
 
     return workflow
 
 
 def init_gifti_surface_wf(name='gifti_surface_wf'):
+    """
+    Extract surfaces from FreeSurfer derivatives folder and
+    re-center GIFTI coordinates to fit align to native T1 space
+
+    """
     workflow = pe.Workflow(name=name)
 
     inputnode = pe.Node(niu.IdentityInterface(['subjects_dir', 'subject_id']), name='inputnode')
@@ -483,56 +484,7 @@ def init_gifti_surface_wf(name='gifti_surface_wf'):
                            name='surface_list', run_without_submitting=True)
     fs_2_gii = pe.MapNode(fs.MRIsConvert(out_datatype='gii'),
                           iterfield='in_file', name='fs_2_gii')
-
-    def normalize_surfs(in_file):
-        """ Re-center GIFTI coordinates to fit align to native T1 space
-
-        For midthickness surfaces, add MidThickness metadata
-
-        Coordinate update based on:
-        https://github.com/Washington-University/workbench/blob/1b79e56/src/Algorithms/AlgorithmSurfaceApplyAffine.cxx#L73-L91
-        and
-        https://github.com/Washington-University/Pipelines/blob/ae69b9a/PostFreeSurfer/scripts/FreeSurfer2CaretConvertAndRegisterNonlinear.sh#L147
-        """
-        import os
-        import numpy as np
-        import nibabel as nib
-        img = nib.load(in_file)
-        pointset = img.get_arrays_from_intent('NIFTI_INTENT_POINTSET')[0]
-        coords = pointset.data
-        c_ras_keys = ('VolGeomC_R', 'VolGeomC_A', 'VolGeomC_S')
-        ras = np.array([float(pointset.metadata[key])
-                        for key in c_ras_keys])
-        # Apply C_RAS translation to coordinates
-        pointset.data = (coords + ras).astype(coords.dtype)
-
-        secondary = nib.gifti.GiftiNVPairs('AnatomicalStructureSecondary',
-                                           'MidThickness')
-        geom_type = nib.gifti.GiftiNVPairs('GeometricType', 'Anatomical')
-        has_ass = has_geo = False
-        for nvpair in pointset.meta.data:
-            # Remove C_RAS translation from metadata to avoid double-dipping in FreeSurfer
-            if nvpair.name in c_ras_keys:
-                nvpair.value = '0.000000'
-            # Check for missing metadata
-            elif nvpair.name == secondary.name:
-                has_ass = True
-            elif nvpair.name == geom_type.name:
-                has_geo = True
-        fname = os.path.basename(in_file)
-        # Update metadata for MidThickness/graymid surfaces
-        if 'midthickness' in fname.lower() or 'graymid' in fname.lower():
-            if not has_ass:
-                pointset.meta.data.insert(1, secondary)
-            if not has_geo:
-                pointset.meta.data.insert(2, geom_type)
-        img.to_filename(fname)
-        return os.path.abspath(fname)
-
-    fix_surfs = pe.MapNode(
-        niu.Function(function=normalize_surfs),
-        iterfield='in_file',
-        name='fix_surfs')
+    fix_surfs = pe.MapNode(NormalizeSurf(), iterfield='in_file', name='fix_surfs')
 
     workflow.connect([
         (inputnode, get_surfaces, [('subjects_dir', 'subjects_dir'),
@@ -551,13 +503,15 @@ def init_gifti_surface_wf(name='gifti_surface_wf'):
         (surface_list, fs_2_gii, [('out', 'in_file')]),
         (fs_2_gii, fix_surfs, [('converted', 'in_file')]),
         (fix_surfs, outputnode, [('out', 'surfaces')]),
-        ])
-
+    ])
     return workflow
 
 
 def init_anat_reports_wf(reportlets_dir, skull_strip_ants, output_spaces,
                          template, freesurfer, name='anat_reports_wf'):
+    """
+    Set up a battery of datasinks to store reports in the right location
+    """
     workflow = pe.Workflow(name=name)
 
     inputnode = pe.Node(
@@ -621,6 +575,9 @@ def init_anat_reports_wf(reportlets_dir, skull_strip_ants, output_spaces,
 
 def init_anat_derivatives_wf(output_dir, output_spaces, template, freesurfer,
                              name='anat_derivatives_wf'):
+    """
+    Set up a battery of datasinks to store derivatives in the right location
+    """
     workflow = pe.Workflow(name=name)
 
     inputnode = pe.Node(
@@ -699,7 +656,7 @@ def init_anat_derivatives_wf(output_dir, output_spaces, template, freesurfer,
                                 ('t1_seg', 'in_file')]),
         (inputnode, ds_t1_tpms, [('source_file', 'source_file'),
                                  ('t1_tpms', 'in_file')]),
-        ])
+    ])
 
     if freesurfer:
         workflow.connect([
@@ -707,7 +664,7 @@ def init_anat_derivatives_wf(output_dir, output_spaces, template, freesurfer,
             (inputnode, ds_surfs, [('source_file', 'source_file'),
                                    ('surfaces', 'in_file')]),
             (name_surfs, ds_surfs, [('out', 'suffix')]),
-            ])
+        ])
     if 'template' in output_spaces:
         workflow.connect([
             (inputnode, ds_t1_mni_warp, [('source_file', 'source_file'),
@@ -720,6 +677,6 @@ def init_anat_derivatives_wf(output_dir, output_spaces, template, freesurfer,
                                      ('mni_seg', 'in_file')]),
             (inputnode, ds_mni_tpms, [('source_file', 'source_file'),
                                       ('mni_tpms', 'in_file')]),
-            ])
+        ])
 
     return workflow
